remove the rest of enzyme
