--- packages/libraries/core/jest.setup.js
@@ -1,6 +1,0 @@
-// @ts-check
-
-const enzyme = require('enzyme');
-const Adapter = require('@wojtekmaj/enzyme-adapter-react-17');
-
-enzyme.configure({ adapter: new Adapter() });
