+++ packages/utils/test-tools/README.md
@@ -1,3 +1,3 @@
 # Test-tools
 
-This is a dev environment only private package designed to aid in testing components via jest and enzyme
+This is a dev environment only private package designed to aid in testing components via jest
