+++ packages/utils/test-tools/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@fluentui-react-native/test-tools",
   "version": "0.1.1",
-  "description": "Tools and mocks for testing components using jest and enzyme",
+  "description": "Tools and mocks for testing components using jest",
   "main": "lib-commonjs/index.js",
   "module": "src/index.ts",
   "typings": "lib/index.d.ts",
@@ -12,7 +12,6 @@
     "clean": "fluentui-scripts clean",
     "depcheck": "fluentui-scripts depcheck",
     "lint": "fluentui-scripts eslint",
-    "test": "fluentui-scripts jest",
     "update-snapshots": "fluentui-scripts jest -u",
     "prettier": "fluentui-scripts prettier",
     "prettier-fix": "fluentui-scripts prettier --fix true"
