+++ scripts/package.json
@@ -28,17 +28,13 @@
     "@react-native-community/cli-platform-ios": "^12.1.1",
     "@react-native/metro-babel-transformer": "^0.73.0",
     "@rnx-kit/jest-preset": "^0.1.14",
-    "@types/enzyme": "^3.10.5",
     "@types/es6-collections": "^0.5.29",
     "@types/es6-promise": "0.0.32",
     "@types/jest": "^29.0.0",
     "@types/node": "^12.11.2",
     "@types/react-test-renderer": "16.9.0",
     "@uifabric/prettier-rules": "^7.0.3",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.5",
     "depcheck": "^1.0.0",
-    "enzyme": "^3.11.0",
-    "enzyme-to-json": "^3.5.0",
     "find-up": "^5.0.0",
     "fs-extra": "^7.0.1",
     "jest": "^29.2.1",
